Made the command handler a listener
